\name{compHclust-internal}

\alias{A.chc}
\alias{RGI.chc}

\title{Internal Complementary Hierarchical Clustering Functions}

\description{Internal functions for the complementary hierarchical
  clustering procedure.} 

\usage{
A.chc(hc)
RGI.chc(x, xhat)
}

\details{These functions are not to be called by the user.}

\author{Gen Nowak \email{gnowak@stanford.edu} and Robert Tibshirani
  \email{tibs@stanford.edu}}

\keyword{internal}